fix for amd64
git-svn-id: http://svn.cern.ch/guest/lhcb/Boole/trunk@57863 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 41e62c1e151ff5de73155f68f22bef5ad58f1a09

+++ Boole/Digi/DigiAlg/src/MergeEventAlg.cpp
@@ -1,4 +1,4 @@
-// $Id: MergeEventAlg.cpp,v 1.11 2006-04-04 14:59:53 cattanem Exp $
+// $Id: MergeEventAlg.cpp,v 1.12 2006-06-26 16:13:29 cattanem Exp $
 #define MERGEEVENTALG_CPP
 // Include files
 
@@ -366,7 +366,7 @@ std::string MergeEventAlg::resetPath( const std::string& oldPath,
   // Check if /Event already in path and reset
   std::string newPath  = oldPath;
   std::string mainPath = "/Event";
-  unsigned int pos = newPath.find(mainPath);
+  size_t pos = newPath.find(mainPath);
   if ( pos != std::string::npos )
   {
     newPath = newPath.erase( pos, mainPath.length()+1 );
